add orientation
file 'orientation_tickets' contains a list of next orientations with which
to test the next kids. With every time a record is saved, the top entry gets
chopped. At the beginning of every new child, orientation is instructed to E.
Orientation also ends up in the record and its filename.